ml_models/model_cls/requirements.txt
fastapi==0.95.0
uvicorn==0.23.0
torch==2.7.0 # PyTorch sürümünü cihazınıza göre seçin, CUDA kullanıyorsanız uyumlu sürüm seçmelisiniz
transformers==4.30.0
azure-storage-blob
requests==2.28.1
pydantic==1.10.5
python-decouple
aimped
numpy==1.24.4
nltk==3.7
ipython
seqeval
xformers